// src/modules/inventory/reports/yarn-store/material-receive-report/material-receive-report-type.ts
export type MaterialReceiveReportType = {
  LOCATION: string;
  COMPANY_ADDRESS: string;
  COMPANY_NAME: string;
  REMARKS: string;
  ADDITIONAL_QTY: number;
  RCV_ID: number;
  LC_NO?: string;
  PR_DTL_NO: number;
  PR_ID: number;
  ITEM_NO: number;
  RECEIVE_QTY: number;
  T_RECEIVE_QTY: number;
  BRAND_NAME?: string;
  ORIGIN_NAME?: string;
  MODEL?: string;
  ITEM_NAME?: string;
  FINAL_WORK_ORDER_QTY: number;
  BILL_NUMBER?: string;
  AUTO_PR_RCV_UNQ_ID: number;
  PURCHASE_ORDER_NO?: string;
  RECEIVE_DATE?: string | null;
  SUPPLIER_NAME?: string;
  MRR_NO?: string;
  GATE_ENTRY_NO?: string;
  IS_POST_ACC: number;
  STORE_ID: number;
  UOM_NAME?: string;
  CUT_OFF: number;
  ACTUAL_PRICE: number;
  CURRENCY_ID: number;
  CURRENCY_RATE: number;
  CHALLAN_DATE?: string | null;
  STORE_NAME?: string;
  PO_NO?: string;
  CURRENCYCODE?: string;
};